requirements.txt
flask==3.1.0
gunicorn==23.0.0
pytesseract==0.3.13
Pillow==11.1.0
opencv-python-headless==4.10.0.84
gspread==6.1.4
google-auth==2.38.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
